Add a 2-piece mount for the RAMBo.
May need a little work if you want to use screws. The holes came out slightly too small for 3mm screws. I ended up just using zip ties to attach the controller.

Mount is made up of 2 pieces, an extender that provides some distance from the 2020 extrusion so the mount clears the x-axis motor.
And a thin mount for the RAMBo controller that connects to that extender.

Making this 2 pieces gives a few advantages:
- The whole thing can be printed flat without any support material (any other orientation would require significant support material due to the nature of this mount.
- The screws connecting the mount to the frame can be accessed without removing the controller from its frame.
- The extender design can be re-used for any other mount as large as the RAMBo which cannot fit the way a RAMPS might.